requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioice==0.9.0
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
av==14.3.0
azure-core==1.33.0
azure-identity==1.21.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cloudevents==1.11.0
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
executing==2.2.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.41
google-crc32c==1.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.10.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.72.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
platformdirs==4.3.7
prance==23.6.21.0
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
semantic-kernel==1.27.2
setuptools==69.0.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
tavily-python==0.5.4
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
